endure::auditor::common
Trait AuditProfileCheck

pub trait AuditProfileCheck {
    // Required method
    fn has_audit_profile(audit_profile: &AuditProfile) -> bool;
}

A trait implemented by the auditors checking if they should be executed for the specified AuditProfile.

Required Methods

fn has_audit_profile(audit_profile: &AuditProfile) -> bool

Checks if the auditor should be run for the specified profile.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.