endure::auditor::common
Trait DHCPv4PacketAuditor

pub trait DHCPv4PacketAuditor:
    Debug
    + Send
    + Sync {
    // Required method
    fn audit(
        &mut self,
        source_ip_address: &Ipv4Addr,
        dest_ip_address: &Ipv4Addr,
        packet: &mut SharedPartiallyParsedPacket,
    );
}

A trait that must be implemented by each DHCPv4 auditor.

The crate::analyzer::Analyzer calls the DHCPv4PacketAuditor::audit function for each received BOOTP packet. The auditor runs specialized checks on the packet and updates its local state and maintained metrics.

Required Methods

fn audit( &mut self, source_ip_address: &Ipv4Addr, dest_ip_address: &Ipv4Addr, packet: &mut SharedPartiallyParsedPacket, )

Runs an audit on the received packet.

The audit is specific to the given auditor implementing this trait. The auditor maintains some specific metrics gathered from the constant analysis of the received packets. It may discard some of the packets that don’t meet the audit criteria.

For example: an auditor checking client retransmissions should ignore the replies from the server and return immediately.

Parameters

source_ip_address - a source IP address of the packet.

dest_ip_address - a destination IP address of the packet.

packet - a partially parsed DHCPv4 or BOOTP packet to be audited