endure::auditor::common
Trait DHCPv4TransactionAuditor

pub trait DHCPv4TransactionAuditor:
    Debug
    + Send
    + Sync {
    // Required method
    fn audit(&mut self, transaction: &mut DHCPv4Transaction);
}

A trait that must be implemented by the transactional DHCPv4 auditors.

The crate::analyzer::Analyzer calls the DHCPv4TransactionAuditor::audit for the groups of related packets (queries and responses) called transactions. Such auditors produce the metrics based on the relationships between the client and server messages. For example: such auditors can measure how long it takes to complete a 4-way exchange.

Required Methods

fn audit(&mut self, transaction: &mut DHCPv4Transaction)

Runs an audit on the transaction.

The audit is specific to the given auditor implementing this trait. The auditor maintains some specific metrics gathered from the constant analysis of the received packets. It may discard some of the transactions that don’t meet the audit criteria.

Parameters

transaction is a DHCPv4 transaction holding client requests and server responses for the particular xid (transaction identifier).