endure::auditor::common
Trait GenericPacketAuditor

pub trait GenericPacketAuditor:
    Debug
    + Send
    + Sync {
    // Required method
    fn audit(&mut self, packet: &PacketWrapper);
}

A trait that must be implemented by auditors running checks on unparsed packets.

An example auditor implementing this trait is the one that gathers timestamps of the received packets (i.e., super::packet_time::PacketTimeAuditor).

Required Methods

fn audit(&mut self, packet: &PacketWrapper)

Runs an audit on the received packet.

The audit is specific to the given auditor implementing this trait. The auditor maintains some specific metrics gathered from the constant analysis of the received packets. It may discard some of the packets that don’t meet the audit criteria.

Parameters

packet - a packet wrapper holding packet metadata.