endure::auditor

Module conversation

conversation is a module implementing auditors maintaining the statistics of unicast, broadcast and relayed messages.

Metrics

A DHCPv4 server can receive three different types of traffic. When it responds to directly connected clients, it receives broadcast messages. When the clients are connected over a relay, it receives unicast messages sent by one or more relay agents. In both cases it also receives unicast messages from the clients renewing their DHCP leases.

Knowing the proportions of these different types of communications can be helpful in diagnosing issues with the server’s environment. For example, a server behind a relay should not receive broadcast traffic. Similarly, a directly connected server should not receive relayed traffic.

Structs

ConversationAuditor 🔒

ConversationStreamAuditor

An auditor maintaining the statistics of the unicast, broadcast and relayed messages.

ConversationTotalAuditor

An auditor maintaining the total statistics of the unicast, broadcast and relayed messages.

Constants

METRIC_INDEX_BROADCAST 🔒

METRIC_INDEX_RELAYED 🔒

METRIC_INDEX_UNICAST 🔒