endure::auditor
Module spectrum

spectrum is a module implementing auditors maintaining statistics of the DHCPv4 traffic types distribution.

The auditors measure the proportions of the different types of DHCPv4 transactions. In particular, it distinguishes between the 4-way exchanges renewals, rebinds, informs and releases. Knowing these proportions can help discovering renewal or rebind storms related to the issues with the DHCPv4 server. It can also help to understand how the lease lifetimes affect the network congestion.

Structs

DHCPv4SpectrumAuditor 🔒

DHCPv4SpectrumStreamAuditor

An auditor counting different kinds of DHCPv4 transactions.

DHCPv4SpectrumTotalAuditor

An auditor counting different kinds of DHCPv4 transactions.

Constants

METRIC_INDEX_ATTEMPTED_DORA 🔒

METRIC_INDEX_ATTEMPTED_INFORM 🔒

METRIC_INDEX_ATTEMPTED_REBIND 🔒

METRIC_INDEX_ATTEMPTED_RELEASE 🔒

METRIC_INDEX_ATTEMPTED_RENEW 🔒