endure::auditor

Module util

util is a module containing tools for calculating metrics.

Structs

MovingRank

A single record held in the MovingRanks container.

MovingRanks

Tracks and holds the highest metrics associated with different clients.

PercentSMA

Simple moving average for calculating percentages of several related metrics.

RingBuffer 🔒

Ring buffer holding metrics samples.

RoundedSMA

A moving average implementation with an arbitrary precision.

RoundedSTA

A total average implementation with an arbitrary precision.

TotalCounter

Collection of related counters with calculable percentages for each counter.

Traits

Average

An interface to the RoundedSMA and RoundedSTA.

FromMetricScope

A trait for constructing the RoundedSMA or RoundedSTA instances from the [MetricScope].

Percentage

An interface to the TotalCounter and PercentSMA to increase a selected counter by 1 and calculate the percentage of the selected metrics compared to total.