endure::auditor::util
Trait Percentage

pub trait Percentage {
    // Required methods
    fn increase(&mut self, metric_index: usize);
    fn percentage(&self, metric_index: usize) -> f64;
}

An interface to the TotalCounter and PercentSMA to increase a selected counter by 1 and calculate the percentage of the selected metrics compared to total.

Required Methods

fn increase(&mut self, metric_index: usize)

Increases a selected metric by 1.

Parameters

metric_index - an index of a metric to increase.

fn percentage(&self, metric_index: usize) -> f64

Returns the percentage of the selected metrics value among all values.

Parameters