endure
Module dispatcher

dispatcher is a module serving as a core of the endure program.

It installs the listeners and receives captured packets from them. The dispatcher will eventually send the received packets to the specialized modules for analysis. In this early version, it merely prints the received packets in the binary form.

Example Usage

The following snippet shows how to initialize and run the dispatcher to capture DHCPv4 traffic on the bridge100 interface:

let mut dispatcher = dispatcher::Dispatcher::new();
let filter = Filter::new().bootp_server_relay();
dispatcher.add_listener("bridge100", filter).expect("listener already added");
dispatcher.http_address = Some("127.0.0.1:8080".to_string());
dispatcher.enable_prometheus = true;
dispatcher.enable_api = true;
dispatcher.csv_output = CsvOutputType::File("./csv_metrics.csv")
dispatcher.dispatch().await.unwrap();

Structs

Dispatcher

Runs the installed listeners until the stop signal occurs.

RegistryWrapper 🔒

Wraps [Registry] instance and registers a custom collector.

api_metrics_handler

An HTTP handler for exposing the metrics via the REST API.

metrics_handler

An HTTP handler for exposing the metrics to Prometheus.

sse_handler

An HTTP handler for exposing server sent events.

Enums

CsvOutputType

Location of the CSV metrics reports.

DispatchError

An enum of errors returned by the Dispatcher::dispatch