endure::proto
Module bootp

bootp is a module providing the BOOTP message parsing capabilities.

The BOOTP protocol has been described in the https://www.rfc-editor.org/rfc/rfc951.html. The protocol allows a client machine to request an IPv4 address assignment from a server by sending a BOOTREQUEST message over UDP. The DHCPv4 protocol was designed on top of the BOOTP, reusing its fields and extending it with the DHCPv4 options carrying additional client configuration. See https://www.ietf.org/rfc/rfc2131.html for the details.

This module should be used for processing the BOOTP messages. The super::dhcp::v4 module should be used for parsing the DHCPv4 messages.

Structs

HAddr

A structure representing a hardware address.

PartiallyParsedState

Partially parsed packet state.

RawState

Raw packet state.

ReceivedPacket

A structure representing a received bootp packet.

Enums

HType

An enum representing hardware types.

OpCode

An enum representing the bootp message types.

Constants

CHADDR_MAX_LEN

chaddr maximum length.

CHADDR_POS

chaddr position.

CIADDR_POS

ciaddr position.

FILE_MAX_LEN

file maximum length.

FILE_POS

file position.

GIADDR_POS

giaddr position.

HLEN_ETHERNET

Ethernet hardware address length (MAC address length).

HLEN_POS

hlen position.

HOPS_POS

hops position.

HTYPE_POS

htype position.

OPCODE_POS

opcode position.

SECS_POS

secs position.

SIADDR_POS

siaddr position.

SNAME_MAX_LEN

sname maximum length.

SNAME_POS

sname position.

UNUSED_POS

unused field position. It is used to carry the flags in DHCPv4.

XID_POS

xid position.

YIADDR_POS

yiaddr position.