endure::proto
Module buffer

buffer is a module that provides mechanics for manipulating the data in the processed packets.

Structs

ClampedNumber

Wraps a number ensuring it is within a desired range.

ReceiveBuffer

Represents a buffer holding received data.

Enums

BufferError

Represents errors returned by the functions processing the buffers.

Type Aliases

Result 🔒