endure::proto::buffer
Struct ReceiveBuffer

pub struct ReceiveBuffer {
    buffer: Vec<u8>,
}

Represents a buffer holding received data.

It provides the functions to read the formatted data from the buffer. It allows for reading the data using a random access index. It means that the data can be accessed selectively, without a need to parse the entire packet. Depending on the packet structure, the data can be read as 1, 2 or 4 byte numbers with a conversion from network to host byte order. The data can also be read as variable length slices or strings.

The caller must know the positions of the desired data fields. The functions reading the data seek to the specified positions. If the specified position or the read data length are out of bounds the BufferError::ReadOutOfBounds error is returned.

Fields
buffer: Vec<u8>
Implementations

impl ReceiveBuffer

pub fn new(data: Vec<u8>) -> ReceiveBuffer

Instantiates a buffer from a data array.

fn read<const N: usize>(&mut self, pos: u32) -> Result<[u8; N], BufferError>

Attempts to read an array of bytes from the buffer.

Parameters

N is a number of bytes to read

pos is a read start position, where 0 is the first byte in the buffer

Result

It returns a result with an array of bytes. If the position is out of bounds or the length spans beyond the end of the buffer it returns the BufferError::ReadOutOfBounds error.

pub fn read_vec(&mut self, pos: u32, len: usize) -> Result<Vec<u8>, BufferError>

Attempts to read a vector of bytes from the buffer.

This function variant reads variable length data from the buffer. If the number of bytes between the pos and the end of the buffer is lower than len, the function reads the remaining bytes until the end of the buffer. It means that the actual data length can be lower than the specified length value.

Parameters

pos is a read start position, where 0 is the first byte in the buffer

len is a number of bytes to read

Result

It returns a vector of bytes that can be shorter than the desired length. If the pos is out of bounds, it returns the BufferError::ReadOutOfBounds error.

pub fn read_null_terminated( &mut self, pos: u32, max_len: usize, ) -> Result<String, BufferError>

Attempts to read from a buffer into a string.

This function is applicable for reading the null-terminated strings from a buffer when the strings also have maximum length specified in the protocol. For example, the sname field in the bootp protocol can hold a null-terminated string that has a maximum length of 64 bytes.

In practice, the strings sometimes lack the null character at the end. In these cases, the function reads all max_len bytes or all bytes until the end of the buffer.

Parameters

pos is a read start position, where 0 is the first byte in the buffer

max_len is a maximum number of bytes to read (if null character is not found)

Result

It returns a string converted from UTF-8. If such the conversion fails it returns the BufferError::ReadUtf8Conversion error. If the pos is out of bounds, it returns the BufferError::ReadOutOfBounds.

pub fn read_u8(&mut self, pos: u32) -> Result<u8, BufferError>

Attempts to read a u8 number from the buffer.

Parameters

pos is a read start position

Result

It returns a byte read from the buffer. If the specified position is out of bounds the BufferError::ReadOutOfBounds is returned.

pub fn read_u16(&mut self, pos: u32) -> Result<u16, BufferError>

Attempts to read a u16 number from the buffer.

Parameters

pos is a read start position, where 0 is the first byte in the buffer

Result

It returns two bytes converted from the network to host byte order. If the specified position is out of bounds the BufferError::ReadOutOfBounds is returned.

pub fn read_u32(&mut self, pos: u32) -> Result<u32, BufferError>

Attempts to read a u32 number from the buffer.

Parameters

pos is a read start position, where 0 is the first byte in the buffer

Result

It returns four bytes converted from the network to host byte order. If the specified position is out of bounds the BufferError::ReadOutOfBounds is returned.

pub fn read_ipv4(&mut self, pos: u32) -> Result<Ipv4Addr, BufferError>

Attempts to read an IPv4 address from the buffer.

Parameters

pos is a read start position, where 0 is the first byte in the buffer

It reads 4 bytes and converts them to an IPv4 address. If the specified position is out of bounds the BufferError::ReadOutOfBounds is returned.