endure::proto::dhcp
Module v4

dhcp::v4 is a module providing DHCPv4 message parsing capabilities.

The DHCPv4 protocol described in https://www.ietf.org/rfc/rfc2131.html has been built on top of the BOOTP protocol. It reuses its message structures extending them with the DHCP options carrying additional configuration data. This module internally calls the crate::proto::bootp module to parse the fixed fields of the DHCPv4 messages.

This module provides functions to parse DHCPv4 options carried in the messages.

Structs

Flags

A structure representing the flags field in DHCP packet.

OptionClientIdentifier

Parsed option 61 (Client Identifier).

OptionMessageType

Parsed option 53 (DHCP Message Type).

PartiallyParsedState

Partially parsed packet state.

RawState

Raw packet state.

ReceivedOption

A structure representing an inbound DHCP option.

ReceivedPacket

A structure representing a received DHCP packet.

Enums

MessageType

An enum representing the DHCP message types.

OptionParseError

Represents errors returned by the functions parsing DHCP options.

Constants

MAGIC_COOKIE_POS

Magic cookie position in the packet.

MESSAGE_TYPE_ACK

DHCPACK message code.

MESSAGE_TYPE_ACTIVELEASEQUERY

DHCPACTIVELEASEQUERY message code.

MESSAGE_TYPE_BULKLEASEQUERY

DHCPBULKLEASEQUERY message code.

MESSAGE_TYPE_DECLINE

DHCPDECLINE message code.

MESSAGE_TYPE_DHCPDISCOVER

DHCPDISCOVER message code.

MESSAGE_TYPE_FORCERENEW

DHCPFORCERENEW message code.

MESSAGE_TYPE_INFORM

DHCPINFORM message code.

MESSAGE_TYPE_LEASEACTIVE

DHCPLEASEACTIVE message code.

MESSAGE_TYPE_LEASEQUERY

DHCPLEASEQUERY message code.

MESSAGE_TYPE_LEASEQUERYDONE

DHCPLEASEQUERYDONE message code.

MESSAGE_TYPE_LEASEQUERYSTATUS

DHCPLEASEQUERYSTATUS message code.

MESSAGE_TYPE_LEASEUNASSIGNED

DHCPLEASEUNASSIGNED message code.

MESSAGE_TYPE_LEASEUNKNOWN

DHCPLEASEUNKNOWN message code.

MESSAGE_TYPE_NAK

DHCPNAK message code.

MESSAGE_TYPE_OFFER

DHCPOFFER message code.

MESSAGE_TYPE_RELEASE

DHCPRELEASE message code.

MESSAGE_TYPE_REQUEST

DHCPREQUEST message code.

MESSAGE_TYPE_TLS

DHCPTLS message code.

OPTIONS_POS

DHCPv4 options position in the packet.

OPTION_CODE_CLIENT_IDENTIFIER

Client Identifier option code.

OPTION_CODE_DHCP_MESSAGE_TYPE

DHCP Message Type option code.

OPTION_CODE_END

End option code.

OPTION_CODE_PAD

Pad option code.

OPTION_CODE_PARAMETER_REQUEST_LIST

Parameter Request List option code.

OPTION_CODE_SERVER_IDENTIFIER

Server Identifier option code.

Type Aliases

PartiallyParsedPacket

A shorthand type for the partially parsed packet.

RawPacket

A shorthand type for the raw packet.

SharedPartiallyParsedPacket

A shareable and lockable instance of the PartiallyParsedPacket.