endure
Module sse

sse is a module implementing a Server Sent Events mechanism.

This implementation is based on the following SSE implementation example.

Structs

Event

Event that can be send over the EventGateway.

EventGateway

Collects the events in the program and passes them to the HTTP server.

EventGatewayState 🔒

Enums

Error

Event serializing and sending errors.

EventType

Specifies a type of an event.