Crate endure

endure is a DHCP diagnostics tool.

Modules

analyzer

analyzer is a module containing the packet analysis and reporting logic.

auditor

auditor is a module containing implementations of the auditors.

cli

cli is a module defining and handling the command line arguments of the endure program.

dispatcher

dispatcher is a module serving as a core of the endure program.

pcap_processor

pcap_processor is a module coordinating processing of the pcap files and presenting the results.

proto

proto is a module holding the packets parsing logic for different protocols.

sse

sse is a module implementing a Server Sent Events mechanism.

Functions

main 🔒