endure
Module auditor

auditor is a module containing implementations of the auditors.

Modules

common

common module contains common declarations for different auditors.

conversation

conversation is a module implementing auditors maintaining the statistics of unicast, broadcast and relayed messages.

metric

metric module contains common declarations related to metrics.

opcode

opcode is a module implementing auditors maintaining statistics of the BOOTP message types.

packet_time

packet_time is a module implementing an auditor tracking last packet timestamp.

retransmission

retransmission is a module implementing auditors maintaining the statistics of DHCP retransmissions.

roundtrip

roundtrip is a module implementing auditors maintaining statistics of how long it takes to complete a DHCP transaction.

spectrum

spectrum is a module implementing auditors maintaining statistics of the DHCPv4 traffic types distribution.

util

util is a module containing tools for calculating metrics.