endure::auditor
Module metric

metric module contains common declarations related to metrics.

Constants

METRIC_BOOTP_CONVERSATION_BROADCAST_COUNT

Total number of broadcast messages.

METRIC_BOOTP_CONVERSATION_BROADCAST_PERCENT

Percentage of broadcast messages.

METRIC_BOOTP_CONVERSATION_RELAYED_COUNT

Total number of relayed messages.

METRIC_BOOTP_CONVERSATION_RELAYED_PERCENT

Percentage of relayed messages.

METRIC_BOOTP_CONVERSATION_UNICAST_COUNT

Total number of unicast not-relayed messages.

METRIC_BOOTP_CONVERSATION_UNICAST_PERCENT

Percentage of unicast not-relayed messages.

METRIC_BOOTP_OPCODE_BOOT_REPLIES_COUNT

Total number of BootReply messages.

METRIC_BOOTP_OPCODE_BOOT_REPLIES_PERCENT

Percentage of the BootReply messages.

METRIC_BOOTP_OPCODE_BOOT_REQUESTS_COUNT

Total number of BootRequest messages.

METRIC_BOOTP_OPCODE_BOOT_REQUESTS_PERCENT

Percentage of the BootRequest messages.

METRIC_BOOTP_OPCODE_INVALID_COUNT

Total number of invalid BOOTP message opcodes.

METRIC_BOOTP_OPCODE_INVALID_PERCENT

Percentage of the invalid BOOTP message opcodes.

METRIC_BOOTP_RETRANSMIT_LONGEST_TRYING_CLIENT

Hardware address of the longest retrying client.

METRIC_BOOTP_RETRANSMIT_PERCENT

Percentage of the retransmissions.

METRIC_BOOTP_RETRANSMIT_SECS_AVG

Average value of the secs field.

METRIC_DHCPV4_ROUNDTRIP_DORA_DO_MILLISECONDS_AVG

Average time in milliseconds to complete a successful Discover/Offer exchange.

METRIC_DHCPV4_ROUNDTRIP_DORA_MILLISECONDS_AVG

Average time in milliseconds to complete a successful 4-way (DORA) exchange.

METRIC_DHCPV4_ROUNDTRIP_DORA_RA_MILLISECONDS_AVG

Average time in milliseconds to complete a successful Request/Ack exchange within a 4-way (DORA) exchange.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_DORA_COUNT

Total number of attempted DORA exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_DORA_PERCENT

Percentage of attempted DORA exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_INFORM_COUNT

Total number of attempted Inform exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_INFORM_PERCENT

Percentage of attempted Inform exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_REBIND_COUNT

Total number of attempted Rebind exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_REBIND_PERCENT

Percentage of attempted Rebind exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_RELEASE_COUNT

Total number of attempted Release messages in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_RELEASE_PERCENT

Percentage of attempted Release messages in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_RENEW_COUNT

Total number of attempted Renew exchanges in all exchanges.

METRIC_DHCPV4_SPECTRUM_ATTEMPTED_RENEW_PERCENT

Percentage of attempted Renew exchanges in all exchanges.

METRIC_PACKET_TIME_DATE_TIME

Timestamp of the last analyzed packet.