endure::auditor
Module opcode

opcode is a module implementing auditors maintaining statistics of the BOOTP message types.

They recognize BootRequest and BootReply messages and maintain a total number and the percentage of each message type in the received packets.

Metrics

Keeping track of the BootRequest and BootReply message types can be useful to detect situations when a DHCP server is unable to keep up with the traffic. Another extreme case is when the are only BootRequest messages and no BootReply. It indicates that the server is down or misconfigured. The OpCodeTotalAuditor tracks the absolute number of messages of a certain type and their percentages in all received packets. The OpCodeStreamAuditor tracks the percentages of the message types in most recent packets.

Structs

OpCodeStreamAuditor

An auditor maintaining the statistics of the BOOTP message types.

OpCodeTotalAuditor

An auditor maintaining the total statistics of the BOOTP message types.

Constants

METRIC_INDEX_BOOT_REPLY 🔒

METRIC_INDEX_BOOT_REQUEST 🔒

METRIC_INDEX_INVALID 🔒