endure
Module cli

cli is a module defining and handling the command line arguments of the endure program.

All new arguments and commands should be specified in this module. The module uses the [clap] crate underneath.

Example Usage

To parse the arguments and run the endure program with these arguments simply do:

Cli::parse().run();

Structs

Cli

A structure holding parsed program arguments.

InterfaceArgs 🔒

PcapReportingFormat 🔒

ReportingArgs 🔒

A group of parameters pertaining to the reporting method selection.

Enums

Commands 🔒

An enum that defines the supported subcommands.

Functions

directory_path_file_parser 🔒

A parser checking if the path of the specified file exists.

directory_path_parser 🔒

A parser checking if the specified path is a directory path.

file_path_parser 🔒

A parser checking if the specified file exists.